FIX issue #13
Need to add a control to check if retpolines are already in the function
slice.
If there, they have been added as direct calls, so it is needed to fix
this circumstance.
If not there, then we are have the previous scenario, where they are
"loc." and the previous flow needs to be executed.

Signed-off-by: Alessandro Carminati <alessandro.carminati@gmail.com>