Implemented support for additional DB engines
This patch includes the changes required to support additional DB
engines, currently MySQL/MariaDB and SQLite.

It also simplifies the configuration, since most of the required
DB related parameters are now supplied as part of the DBDSN
configuration setting (-d command line option), that should be
supplied in the DB engine (selected with the -e command line
option) specific format.

Signed-off-by: Lev Veyde <lveyde@gmail.com>