Fixed bug that was causing panic on missing ENABLE_VERSION_CONFIG
In case the utility was running with Mode that is missing the
ENABLE_VERSION_CONFIG flag, the utility would panic.

That happens due to a race condition, since in that case the
Instance ID, which is later used, was not properly allocated.

This patch fixes the issue, making sure the Instance ID is
always allocated.

Signed-off-by: Lev Veyde <lveyde@gmail.com>