Refactor Nav to Include Graphviz Functionality for Diagram Display
This commit introduces an enhancement to the Nav commandline tool that
integrates Graphviz functionality for generating diagrams within the
tool itself, thereby eliminating the need to pipe output to an external
Graphviz tool like dot. Although there is a long-term goal of creating
a web application for navigating code with diagram display, the objective
for now is to maintain Nav as a commandline tool.

Signed-off-by: Alessandro Carminati <alessandro.carminati@gmail.com>