Mock Driver for Persistence-Related Functions
This patch introduces a new mock driver for the persistence-related functions
that are used in the db version. The primary objective of this patch is to
remove the coupling between high-level function tests and the persistence
layer in the long term. By replacing the functions that depend on the
persistence layer with a mock driver, we can test the high-level functions
without actually relying on the persistence layer.
This also means that we can easily switch between different persistence
implementations in the future without having to update our tests.

Signed-off-by: Alessandro Carminati <alessandro.carminati@gmail.com>
Suggested-by: Maurizio Papini <mpapini@redhat.com>