Container: dependecies updated
This commit reviews issue #21, where a user reported a failure to build the
container. The issue stemmed from changes in the `capstone` dependency of
Radare2, resulting in build failures. Additionally, recent cgo builds seem
to have specific dependencies with glibc, rendering the Alpine base
unsuitable for the project, which is based on musl. This commit primarily
updates the Radare2 version and switches the base from Alpine to CS9.

Signed-off-by: Alessandro Carminati <alessandro.carminati@gmail.com>