gdata display: Adds Global data handling
This commit adds to ks-nav the capability to handle a new visualization mode
that displays global data usage in kernel functions.

Both the kern-bin_db and nav utilities are impacted by this modification.
kern-bin_db now uses the nm utility to collect data symbols, and the database
 schema has two additional tables. The nav utility implements the logic to
display the new visualization mode.

Signed-off-by: Alessandro Carminati <alessandro.carminati@gmail.com>