Enhance Kernel Documentation with Local and Global Symbol Differentiation
This commit addresses discussions from the Elisa Architecture WG regarding
the expansion of kernel documentation. It was highlighted that kernel
functions local to the compiler units were not a priority for consideration
by the working group. Currently, ks-nav nm-symbols data are stored using
the same identifier for both local and global symbols.
This update introduces a mechanism to differentiate between local and
global text symbols in new databases.

Signed-off-by: Alessandro Carminati <alessandro.carminati@gmail.com>